correction bug artisan:up

+++ updates/add_waka_mails_table_u102.php
@@ -4,26 +4,31 @@ use October\Rain\Database\Schema\Blueprint;
 use October\Rain\Database\Updates\Migration;
 use Schema;
 
-class CreateWakaMailsTableU102 extends Migration
+class AddWakaMailsTableU102 extends Migration
 {
     public function up()
     {
-        Schema::table('waka_mailer_wakamails', function (Blueprint $table) {
-            $table->boolean('use_lp')->nullable();
-            $table->string('lp')->nullable();
-            $table->boolean('use_key')->nullable();
-            $table->string('key_duration')->nullable();
+        if (Schema::hasTable('waka_mailer_wakamails')) {
+            Schema::table('waka_mailer_wakamails', function (Blueprint $table) {
+                $table->boolean('use_lp')->nullable();
+                $table->string('lp')->nullable();
+                $table->boolean('use_key')->nullable();
+                $table->string('key_duration')->nullable();
 
-        });
+            });
+        }
+
     }
 
     public function down()
     {
-        Schema::table('waka_mailer_wakamails', function (Blueprint $table) {
-            $table->dropColumn('use_lp');
-            $table->dropColumn('lp');
-            $table->dropColumn('use_key');
-            $table->dropColumn('key_duration');
-        });
+        if (Schema::hasTable('waka_mailer_wakamails')) {
+            Schema::table('waka_mailer_wakamails', function (Blueprint $table) {
+                $table->dropColumn('use_lp');
+                $table->dropColumn('lp');
+                $table->dropColumn('use_key');
+                $table->dropColumn('key_duration');
+            });
+        }
     }
 }
