+++ updates/add_waka_mailtos_table_U103.php
@@ -8,22 +8,26 @@ class AddWakaMailtosTableU103 extends Migration
 {
     public function up()
     {
-        Schema::table('waka_mailtoer_waka_mailtos', function (Blueprint $table) {
-            $table->boolean('use_lp')->nullable();
-            $table->string('lp')->nullable();
-            $table->boolean('use_key')->nullable();
-            $table->string('key_duration')->nullable();
+        if (Schema::hasTable('waka_mailtoer_waka_mailtos')) {
+            Schema::table('waka_mailtoer_waka_mailtos', function (Blueprint $table) {
+                $table->boolean('use_lp')->nullable();
+                $table->string('lp')->nullable();
+                $table->boolean('use_key')->nullable();
+                $table->string('key_duration')->nullable();
 
-        });
+            });
+        }
     }
 
     public function down()
     {
-        Schema::table('waka_mailtoer_waka_mailtos', function (Blueprint $table) {
-            $table->dropColumn('use_lp');
-            $table->dropColumn('lp');
-            $table->dropColumn('use_key');
-            $table->dropColumn('key_duration');
-        });
+        if (Schema::hasTable('waka_mailtoer_waka_mailtos')) {
+            Schema::table('waka_mailtoer_waka_mailtos', function (Blueprint $table) {
+                $table->dropColumn('use_lp');
+                $table->dropColumn('lp');
+                $table->dropColumn('use_key');
+                $table->dropColumn('key_duration');
+            });
+        }
     }
 }
